upgrade parent and core versions

+++ src/main/java/io/github/davemeier82/homeautomation/shelly/device/property/ShellyDimmerRelay.java
@@ -44,7 +44,7 @@ public class ShellyDimmerRelay extends AbstractDimmerRelay {
    * @param eventFactory   the event factory
    * @param mqttClient     the MQTT client
    */
-  public ShellyDimmerRelay(long id,
+  public ShellyDimmerRelay(int id,
                            Device device,
                            String relayTopic,
                            String dimmerTopic,
