+++ src/main/java/io/github/davemeier82/homeautomation/shelly/device/property/ShellyRoller.java
@@ -43,7 +43,7 @@ public class ShellyRoller extends AbstractRoller {
    * @param eventFactory   the event factory
    * @param mqttClient     the MQTT client
    */
-  public ShellyRoller(long id,
+  public ShellyRoller(int id,
                       Device device,
                       String topic,
                       EventPublisher eventPublisher,
